requirements.txt
flask==2.3.3
openai==1.12.0
PyPDF2==3.0.1
python-docx==1.0.1
werkzeug==2.3.7
anthropic==0.37.1